Made it possible to mark adjacent words
When a full word is selected the adjacent words (those to possibly navigating to
next) is marked. It does not mark across wrap however. This is because when at
the end of the buffer for instance it is confusing two see the two previous
words marked (because the next wraps to the first). It is also possibly to
control whether to mark words or not through a setting